add correct dylib for macOS vosk
