+++ app/src/main/java/logic/Platform.java
@@ -1,0 +1,5 @@
+package logic;
+
+public enum Platform {
+    WINDOWS, LINUX, MACOS
+}
